moved login header to common folder
