+++ src/components/Login/signup.js
@@ -7,7 +7,7 @@ import TextField from "@material-ui/core/TextField";
 import { withStyles } from "@material-ui/styles";
 
 // internal components
-import Header from "./header.js";
+import Header from "../common/LoginOnboardingHeader/index.js";
 import StyledButton from "../common/StyledButton/index.js";
 import ErrorMsgBox from "../common/ErrorMsgBox/index.js";
 
@@ -259,7 +259,6 @@ class SignUp extends Component {
               <StyledButton
                 btnText="github"
                 type="black"
-                hasIcon={true}
                 iconClass="fab fa-github"
               />
               <span className={classes.horizontalLine}>or</span>
